Setup strategy protocol for ball-ball collision
